pyproject.toml
[tool.poetry]
name = "project2"
version = "0.1.0"
description = ""
authors = ["jinwoo <sds7629>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
environ = "^1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.4"
django-debug-toolbar = "^4.2.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.2.0"
pymysql = "^1.1.0"
gunicorn = "^21.2.0"
django-environ = "^0.11.2"
mysqlclient = "^2.2.0"
django-filter = "^23.2"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
